LPS-79388 (message-boards-uad-test) Rename UADEntityTestHelper to UADTestHelper
