+++ modules/apps/collaboration/message-boards/message-boards-uad-test/src/testIntegration/java/com/liferay/message/boards/uad/display/test/MBMessageUADDisplayTest.java
@@ -18,7 +18,7 @@ import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 
 import com.liferay.message.boards.model.MBMessage;
 import com.liferay.message.boards.uad.constants.MBUADConstants;
-import com.liferay.message.boards.uad.test.MBMessageUADEntityTestHelper;
+import com.liferay.message.boards.uad.test.MBMessageUADTestHelper;
 
 import com.liferay.portal.kernel.model.BaseModel;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
@@ -51,7 +51,7 @@ public class MBMessageUADDisplayTest extends BaseUADDisplayTestCase {
 
 	@Override
 	protected BaseModel<?> addBaseModel(long userId) throws Exception {
-		MBMessage mbMessage = _mbMessageUADEntityTestHelper.addMBMessage(userId);
+		MBMessage mbMessage = _mbMessageUADTestHelper.addMBMessage(userId);
 
 		_mbMessages.add(mbMessage);
 
@@ -80,13 +80,13 @@ public class MBMessageUADDisplayTest extends BaseUADDisplayTestCase {
 
 	@After
 	public void tearDown() throws Exception {
-		_mbMessageUADEntityTestHelper.cleanUpDependencies(_mbMessages);
+		_mbMessageUADTestHelper.cleanUpDependencies(_mbMessages);
 	}
 
 	@DeleteAfterTestRun
 	private final List<MBMessage> _mbMessages = new ArrayList<MBMessage>();
 	@Inject
-	private MBMessageUADEntityTestHelper _mbMessageUADEntityTestHelper;
+	private MBMessageUADTestHelper _mbMessageUADTestHelper;
 	@Inject(filter = "model.class.name=" +
 	MBUADConstants.CLASS_NAME_MB_MESSAGE)
 	private UADAggregator _uadAggregator;
